+++ pyvelm/modules/geo_data/models/__init__.py
@@ -1,0 +1,4 @@
+from . import continent  # noqa: F401
+from . import country  # noqa: F401
+from . import state  # noqa: F401
+from . import city  # noqa: F401